Replace radix tree with a hash map
This change simplifies the code significantly. The main advantage of
using a radix tree was that it would allow for finding all the entries
whose names start with a common prefix, which could be used for finding
the descendants of a directory if the archive were used for storing
files.