Remove `OpenStore` interface
This commit remove the `OpenStore` trait in favor of allowing each
data store to provide its own methods for creating new instances. The
rationale for this change is that the behavior of each data store that
implemented the trait was slightly different, and they were not
consistent in the circumstances that caused them to return the
documented error types. This is because the different underlying storage
mechanisms are too different in their behavior to encapsulate in a
single trait.

This commit also simplifies the code for opening/creating data stores so
that now all data stores will open the store if it exists or create it
if it does not. The user still has control over how repositories are
opened/created through the `OpenOptions` type, but now we don't have to
re-implement this behavior for each data store. Most of the logic in
each data store was just handling the different options provided to
`OpenStore::open`, which resulted in a lot of redundant code.

Finally, this commit adds the ability to specify a path prefix in the
`S3Store` data store.