Remove file-based repository locking
Previously, file locks were used to lock repositories between processes
within the same machine. This commit removes that feature so that
repositories are only locked within the same process. The file-based
locking system had too many inconsistencies, like that it only worked
between processes run by the same user and that behavior was different
depending on whether the repository was already open in the same process
or in a different process.

This commit also removes the ability to block until the repository is
unlocked, as that was only available for file-based locking.