Remove associated type from `DataStore` trait
Considering that the type of the errors returned by the data store are
erased in `crate::Error::Store`, there is little reason to preserve this
type information as an associated type in the `DataStore` interface.
This commit removes the associated type and has data stores return a
boxed error directly.

This change makes `DataStore` object-safe, allowing us to remove a type
parameter from repository types.