Remove store type parameter from repository types
Previously, each repository type had a type parameter for the type of
data store it contained. With this commit, the wrapped data store is
boxed and type-erased, allowing us to remove the type parameter.

This type parameter made some parts of the API more clunky, specifically
opening and creating repositories. It was also unnecessary. While it is
no longer possible to get back the unboxed data store which was passed
in, the only real use-case for this feature to begin with was for
closing and then re-opening a repository.