Remove creation time from repo metadata
Previously, the repository stored the time it was created in its
unencrypted metadata. This data wasn't being used anywhere within the
library, and there isn't a convincing reason to keep this information
available unencrypted, so it's probably best to leave it to the user to
store this information in the repository themselves if they need access
to it.