Don't implicitly clean when committing
This commit alters the way repositories commit data so that they don't
implicitly attempt to clean themselves after the commit. The rationale
for this change is that cleaning can be an expensive operation,
especially when packing is enabled, and the user should have more
control over when they clean the repository.