Store headers for savepoints in repository
This commit changes the way savepoints work so that they no longer store
the repository header in memory.

Advantages:
- Less memory usage. Previously, adding a savepoint would approximately
double the memory usage of the repository. Also, the memory usage would
scale with the number of savepoints.
- Easy to clone savepoints. Previously, savepoints could not be easily
cloned because the memory usage would scale with the number of clones.

Disadvantages:
- Savepoints can invalidate other savepoints.
- Creating and restoring savepoints requires accessing the data store.
- Committing changes is more complex.