Replace `ObjectRepo` with `KeyRepo`
This commit removes `ObjectRepo` and replaces it with `KeyRepo` as the
base repository type which other repository types are implemented on top
of. The API of `ObjectRepo` was needlessly complicated, and while it was
designed to give users more control over the tradeoff between
performance and memory usage, the memory which could be saved using
this API didn't amount to much except in exceptional cases. This change
allowed for performing some optimizations which should reduce memory
usage for repositories with a large number of objects. This change also
should also allow for fixing a major bug which caused rolling back
changes and restoring to a savepoint to not work properly in
`ObjectRepo`.