Remove lifetime parameter on `Restore`
Previously, a phantom lifetime parameter was used to ensure that a
`Restore` is never completed on a different repository instance from the
one which it was started on. This commit changes `Restore` to remove the
lifetime parameter and instead store the instance ID to track the
repository instance it's associated with. This simplifies the type by
removing the lifetime, allows for the `Restore` to be completed if the
user switches instances and then later switches back, and avoids relying
on the fact that switching instances drops the repository value.