Disable automatic flush on drop in objects
Previously, objects automatically attempted to flush when they were
dropped. The intent behind this decision was to mimic the behavior of
`std::fs::File`. Changing the behavior so objects no longer
automatically flush allows for aborting failed writes by dropping the
object. Since no data is persisted to the repository until a successful
flush, this allows for objects to be used transactionally.