Make `NoSpecial` and `NoMetadata` uninstantiable
Previously, `NoSpecial` and `NoMetadata` were unit structs, which could
lead to the confusing situation of having a special entry of type
`EntryType::Special(NoSpecial)` or an entry with metadata
`Some(NoMetadata)`. It is not clear what these values represent. Now,
`NoSpecial` and `NoMetadata` are zero-variant enums so they cannot be
instantiated, and `FileMetadata` has been updated so it can return
`None`.