Fix bugs with `ObjectId`
Previously, `ObjectId` contained the instance ID of the object it is
associated with. However, this was unnecessary because object handle IDs
are unique between all instances of the same repository.

Also, it was previously possible to get the `ObjectId` of an object even
when the object had been invalidated. This was a problem because object handle
IDs can be reused once an object is removed.